+++ kubernetes/workspace_controller/services/pool_service.py
@@ -1671,165 +1671,4 @@ fi
 echo "Feature installation completed successfully"
 """
 
-    def _create_wrapper_dockerfile_script(self, workspace_ids: dict, workspace_config: dict) -> str:
-        """Generate script to create a wrapper Dockerfile using the user's Dockerfile as a base."""
-        repo_name = workspace_config['repo_name']
-        init_script = f"""
-        # Create directory for wrapper Dockerfile and user Dockerfile
-        mkdir -p /workspaces/.code-server-wrapper
-        mkdir -p /workspaces/.user-dockerfile
-        cd /workspaces/.code-server-wrapper
-
-        # Locate the user's Dockerfile in their repo
-        USER_REPO_PATH="/workspaces/{repo_name}"
-        DOCKERFILE_PATH="$USER_REPO_PATH/.devcontainer/Dockerfile"
-        DEVCONTAINER_JSON_PATH="$USER_REPO_PATH/.devcontainer/devcontainer.json"
-
-        # Debugging and validation
-        echo "DEBUG: Checking repository and Dockerfile"
-        if [ -d "$USER_REPO_PATH" ]; then
-            echo "DEBUG: Repository directory exists at $USER_REPO_PATH"
-            ls -la "$USER_REPO_PATH"
-        else
-            echo "DEBUG: ERROR - Repository directory does not exist at $USER_REPO_PATH"
-        fi
-
-        if [ -d "$USER_REPO_PATH/.devcontainer" ]; then
-            echo "DEBUG: .devcontainer directory exists"
-            ls -la "$USER_REPO_PATH/.devcontainer"
-        else
-            echo "DEBUG: .devcontainer directory does not exist"
-        fi
-
-        if [ -f "$DOCKERFILE_PATH" ]; then
-            echo "DEBUG: Dockerfile exists at $DOCKERFILE_PATH"
-            cat "$DOCKERFILE_PATH" | head -n 10
-        else
-            echo "DEBUG: Dockerfile does not exist at $DOCKERFILE_PATH"
-        fi
-
-        if [ -f "$DEVCONTAINER_JSON_PATH" ]; then
-            echo "DEBUG: devcontainer.json exists at $DEVCONTAINER_JSON_PATH"
-            cat "$DEVCONTAINER_JSON_PATH" | head -n 20
-        else
-            echo "DEBUG: devcontainer.json does not exist at $DEVCONTAINER_JSON_PATH"
-        fi
-
-        # Clone repository if needed
-        if [ ! -d "$USER_REPO_PATH" ]; then
-            echo "Repository not found at $USER_REPO_PATH, attempting to clone again"
-            cd /workspaces
-
-            BRANCH="{workspace_config['github_branches'][0] if workspace_config['github_branches'] and workspace_config['github_branches'][0] else ''}"
-
-            if [ ! -z "$BRANCH" ]; then
-                echo "Cloning with specific branch: $BRANCH"
-                git clone -b $BRANCH {workspace_config['github_urls'][0]} {repo_name}
-            else
-                echo "Cloning with default branch"
-                git clone {workspace_config['github_urls'][0]} {repo_name}
-            fi
-
-            git config --global --add safe.directory /workspaces/{repo_name}
-        fi
-
-        # Check again after potential re-cloning
-        if [ -f "$DOCKERFILE_PATH" ]; then
-            echo "Found user Dockerfile at $DOCKERFILE_PATH"
-            cp "$DOCKERFILE_PATH" /workspaces/.user-dockerfile/Dockerfile
-
-            if [ -f "$DEVCONTAINER_JSON_PATH" ]; then
-                echo "Found devcontainer.json - processing configuration"
-                cp "$DEVCONTAINER_JSON_PATH" /workspaces/.user-dockerfile/
-
-                if ! command -v jq &> /dev/null; then
-                    echo "Installing jq to parse devcontainer.json"
-                    apt-get update && apt-get install -y jq tmux
-                fi
-
-                EXTENSIONS=$(jq -r '.extensions[]? // empty' "$DEVCONTAINER_JSON_PATH" 2>/dev/null)
-                if [ -z "$EXTENSIONS" ]; then
-                    EXTENSIONS=$(jq -r '.customizations.vscode.extensions[]? // empty' "$DEVCONTAINER_JSON_PATH" 2>/dev/null)
-                fi
-
-                if [ ! -z "$EXTENSIONS" ]; then
-                    echo "$EXTENSIONS" > /workspaces/.extensions-list
-                fi
-
-                SETTINGS=$(jq -r '.settings // .customizations.vscode.settings // empty' "$DEVCONTAINER_JSON_PATH" 2>/dev/null)
-                if [ ! -z "$SETTINGS" ]; then
-                    mkdir -p /workspaces/.vscode
-                    echo "$SETTINGS" > /workspaces/.vscode/settings.json
-                fi
-
-                FEATURES=$(jq -r '.features // empty' "$DEVCONTAINER_JSON_PATH" 2>/dev/null)
-                if [ ! -z "$FEATURES" ]; then
-                    echo "$FEATURES" > /workspaces/.devcontainer-features
-                fi
-
-                PORTS=$(jq -r '.forwardPorts[]? // empty' "$DEVCONTAINER_JSON_PATH" 2>/dev/null)
-                if [ ! -z "$PORTS" ]; then
-                    echo "$PORTS" > /workspaces/.forward-ports
-                fi
-
-                CUSTOMIZATIONS=$(jq -r '.customizations // empty' "$DEVCONTAINER_JSON_PATH" 2>/dev/null)
-                if [ ! -z "$CUSTOMIZATIONS" ]; then
-                    echo "$CUSTOMIZATIONS" > /workspaces/.customizations
-                fi
-
-                ENV_VARS=$(jq -r '.containerEnv // empty | to_entries[] | "\(.key)=\(.value)"' "$DEVCONTAINER_JSON_PATH" 2>/dev/null)
-                if [ ! -z "$ENV_VARS" ]; then
-                    echo "$ENV_VARS" > /workspaces/.container-env
-                fi
-
-                REMOTE_ENV_VARS=$(jq -r '.remoteEnv // empty | to_entries[] | "\(.key)=\(.value)"' "$DEVCONTAINER_JSON_PATH" 2>/dev/null)
-                if [ ! -z "$REMOTE_ENV_VARS" ]; then
-                    echo "$REMOTE_ENV_VARS" > /workspaces/.remote-env
-                fi
-
-                REMOTE_USER=$(jq -r '.remoteUser // empty' "$DEVCONTAINER_JSON_PATH" 2>/dev/null)
-                CONTAINER_USER=$(jq -r '.containerUser // empty' "$DEVCONTAINER_JSON_PATH" 2>/dev/null)
-
-                if [ ! -z "$REMOTE_USER" ]; then
-                    echo "REMOTE_USER=$REMOTE_USER" > /workspaces/.user-config
-                fi
-
-                if [ ! -z "$CONTAINER_USER" ]; then
-                    echo "CONTAINER_USER=$CONTAINER_USER" >> /workspaces/.user-config
-                fi
-
-                POST_CREATE_CMD=$(jq -r '.postCreateCommand // empty' "$DEVCONTAINER_JSON_PATH" 2>/dev/null)
-                if [ ! -z "$POST_CREATE_CMD" ]; then
-                    echo "$POST_CREATE_CMD" > /workspaces/post-create-command.sh
-                    chmod +x /workspaces/post-create-command.sh
-                fi
-
-                POST_START_CMD=$(jq -r '.postStartCommand // empty' "$DEVCONTAINER_JSON_PATH" 2>/dev/null)
-                if [ ! -z "$POST_START_CMD" ]; then
-                    echo "$POST_START_CMD" > /workspaces/post-start-command.sh
-                    chmod +x /workspaces/post-start-command.sh
-                fi
-            fi
-        else
-            echo "Warning: No Dockerfile found at $DOCKERFILE_PATH"
-            echo "Using default Go dev container image instead"
-            echo "FROM mcr.microsoft.com/devcontainers/go:latest" > /workspaces/.user-dockerfile/Dockerfile
-        fi
-        """
-
-    def _create_wrapper_dockerfile(self, workspace_ids: dict) -> str:
-        """Create a wrapper Dockerfile using the user's image as a base."""
-        return f"""
-        cat > Dockerfile << 'EOF'
-        FROM {self.aws_account_id}.dkr.ecr.us-east-1.amazonaws.com/workspace-images:custom-user-{workspace_ids['namespace_name']}-{workspace_ids['build_timestamp']}
-
-        RUN git config --global --add safe.directory /workspaces && \
-            git config --global --add safe.directory '*'
-
-        RUN curl -fsSL https://code-server.dev/install.sh | sh
-
-        EXPOSE 8444
-
-        ENTRYPOINT ["/bin/bash", "-c", "if [ -f /workspaces/install-features.sh ]; then /workspaces/install-features.sh; fi && if [ -f /workspaces/setup-env.sh ]; then source /workspaces/setup-env.sh; fi && if [ -f /workspaces/install-extensions.sh ]; then /workspaces/install-extensions.sh; fi && if [ -f /workspaces/run-lifecycle.sh ]; then /workspaces/run-lifecycle.sh & fi && /usr/bin/code-server --bind-addr 0.0.0.0:8444 --auth password --user-data-dir /config/data --extensions-dir /config/extensions /workspaces"]
-        EOF
-        """
+
